+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: dagutin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/03 13:07:22 by dagutin           #+#    #+#             */
-/*   Updated: 2022/10/03 13:07:25 by dagutin          ###   ########.fr       */
+/*   Updated: 2022/10/03 13:31:39 by dagutin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
